Clean up a bit and separate debugger in 2 files
